requirements.txt
gunicorn==20.0.4
numpy>=1.22.0
opencv-python==4.5.3.56
Pillow==8.3.2
pytesseract==0.3.3
Flask==1.1.1
opencv-contrib-python==4.5.3.56
Flask-RESTful==0.3.8
flask-mysqldb==1.0.1
flask-migrate==3.0.1
bcrypt==3.2.0
werkzeug==2.0.3
flask-cors==3.1.1
scikit-learn==0.24.2
pandas==1.2.4
flask-wtf==0.14.3
flask-login==0.5.0
flask-sqlalchemy==2.5.1
Flask-Bcrypt==1.0.1